+++ tests/test_errors.py
@@ -11,7 +11,7 @@ from wemake_python_styleguide.errors.base import (
 def test_visitor_returns_location():
     """Ensures that `BaseNodeVisitor` return correct error message."""
     visitor = ASTStyleViolation(node=ast.parse(''), text='error')
-    visitor.error_template = '{0} {1}'
+    visitor.error_template = '{0}'
     visitor.code = 1
     assert visitor.node_items() == (0, 0, 'Z001 error')