+++ wemake_python_styleguide/visitors/presets/complexity.py
@@ -5,6 +5,7 @@ from wemake_python_styleguide.visitors.ast.complexity import (
     function,
     jones,
     nested,
+    numbers,
     offset,
 )
 
@@ -13,6 +14,7 @@ COMPLEXITY_PRESET = (
     function.FunctionComplexityVisitor,
     jones.JonesComplexityVisitor,
     nested.NestedComplexityVisitor,
+    numbers.MagicNumberVisitor,
     offset.OffsetVisitor,
     counts.ImportMembersVisitor,
     counts.ModuleMembersVisitor,